+++ tests/test_event_handlers.py
@@ -3,7 +3,7 @@ dj-stripe Event Handler tests
 """
 import decimal
 from copy import deepcopy
-from unittest.mock import patch
+from unittest.mock import ANY, call, patch
 
 from django.contrib.auth import get_user_model
 from django.test import TestCase
@@ -25,6 +25,7 @@ from djstripe.models import (
 )
 
 from . import (
+    FAKE_ACCOUNT,
     FAKE_BALANCE_TRANSACTION,
     FAKE_CARD,
     FAKE_CARD_AS_PAYMENT_METHOD,
@@ -51,16 +52,21 @@ from . import (
     FAKE_EVENT_INVOICE_UPCOMING,
     FAKE_EVENT_INVOICEITEM_CREATED,
     FAKE_EVENT_INVOICEITEM_DELETED,
+    FAKE_EVENT_PAYMENT_INTENT_SUCCEEDED_DESTINATION_CHARGE,
     FAKE_EVENT_PAYMENT_METHOD_ATTACHED,
     FAKE_EVENT_PLAN_CREATED,
     FAKE_EVENT_PLAN_DELETED,
     FAKE_EVENT_PLAN_REQUEST_IS_OBJECT,
     FAKE_EVENT_TRANSFER_CREATED,
     FAKE_EVENT_TRANSFER_DELETED,
+    FAKE_FILEUPLOAD_ICON,
+    FAKE_FILEUPLOAD_LOGO,
     FAKE_INVOICE,
     FAKE_INVOICE_II,
     FAKE_INVOICEITEM,
+    FAKE_PAYMENT_INTENT_DESTINATION_CHARGE,
     FAKE_PAYMENT_INTENT_I,
+    FAKE_PAYMENT_METHOD_I,
     FAKE_PLAN,
     FAKE_PRODUCT,
     FAKE_SUBSCRIPTION,
@@ -829,3 +835,66 @@ class TestTransferEvents(EventTestCase):
 
         event = self._create_event(FAKE_EVENT_TRANSFER_DELETED)
         event.invoke_webhook_handlers()
+
+
+class TestPaymentIntentEvents(EventTestCase):
+    """Test case for payment intent event handling."""
+
+    @patch(
+        "stripe.Customer.retrieve", return_value=deepcopy(FAKE_CUSTOMER), autospec=True
+    )
+    @patch(
+        "stripe.Account.retrieve",
+        return_value=deepcopy(FAKE_ACCOUNT),
+        autospec=IS_STATICMETHOD_AUTOSPEC_SUPPORTED,
+    )
+    @patch(
+        "stripe.FileUpload.retrieve",
+        side_effect=(deepcopy(FAKE_FILEUPLOAD_ICON), deepcopy(FAKE_FILEUPLOAD_LOGO)),
+        autospec=True,
+    )
+    @patch(
+        "stripe.PaymentIntent.retrieve",
+        return_value=deepcopy(FAKE_PAYMENT_INTENT_DESTINATION_CHARGE),
+        autospec=True,
+    )
+    @patch(
+        "stripe.PaymentMethod.retrieve",
+        return_value=deepcopy(FAKE_PAYMENT_METHOD_I),
+        autospec=True,
+    )
+    def test_payment_intent_succeeded_with_destination_charge(
+        self,
+        customer_retrieve_mock,
+        account_retrieve_mock,
+        file_upload_retrieve_mock,
+        payment_intent_retrieve_mock,
+        payment_method_retrieve_mock,
+    ):
+        """Test that the payment intent succeeded event can create all related objects.
+
+        This should exercise the machinery to set `stripe_account` when recursing into
+        objects related to a connect `Account`.
+        """
+        event = self._create_event(
+            FAKE_EVENT_PAYMENT_INTENT_SUCCEEDED_DESTINATION_CHARGE
+        )
+        event.invoke_webhook_handlers()
+
+        # Make sure the file uploads were retrieved using the account ID.
+        file_upload_retrieve_mock.assert_has_calls(
+            (
+                call(
+                    id=FAKE_FILEUPLOAD_ICON["id"],
+                    api_key=ANY,
+                    expand=ANY,
+                    stripe_account=FAKE_ACCOUNT["id"],
+                ),
+                call(
+                    id=FAKE_FILEUPLOAD_LOGO["id"],
+                    api_key=ANY,
+                    expand=ANY,
+                    stripe_account=FAKE_ACCOUNT["id"],
+                ),
+            )
+        )